処理を切り離した
